+++ NEWS.md
@@ -13,6 +13,7 @@
 - Fixed problems creating new files and projects on a UNC path (#14963, #14964; Windows Desktop)
 - Prevent attempting to start Copilot on a non-main thread (#14952)
 - Reformat Code no longer inserts whitespace around '^' operator (#14973)
+- Prompt for personal access token instead of password when using github via https (#14103)
 
 #### Posit Workbench
 -